feat(drive): re-enable proof logic
